Move render_evaluation_parameter_string out of util module
